pyproject.toml
[project]
name = "freelance-backend"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.14,<4.0"
dependencies = [
    "requests (>=2.32.5,<3.0.0)",
    "fastapi[standard] (>=0.124.0,<0.125.0)",
    "sqlalchemy (>=2.0.44,<3.0.0)",
    "aiosqlite (>=0.21.0,<0.22.0)",
    "pydantic[email] (>=2.12.5,<3.0.0)",
    "passlib (>=1.7.4,<2.0.0)",
    "loguru (>=0.7.3,<0.8.0)",
    "bcrypt (>=5.0.0,<6.0.0)",
    "celery (>=5.6.0,<6.0.0)",
    "jinja2 (>=3.1.6,<4.0.0)",
    "redis (>=7.1.0,<8.0.0)",
    "billiard (>=4.2.4,<5.0.0)",
    "flower (>=2.0.1,<3.0.0)",
    "eventlet (>=0.40.4,<0.41.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "fastapi-cache2 (>=0.2.2,<0.3.0)",
    "slowapi (>=0.1.9,<0.2.0)"
]

[tool.poetry]
packages = [{include = "freelance_backend", from = "src"}]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "python-dotenv (>=1.2.1,<2.0.0)",
    "alembic (>=1.17.2,<2.0.0)"
]
